Stop using ulong in the CPU core; remove two shifts into bit 31
ulong is uint64_t on native builds but the 32-bit unsigned long from
sys/types.h on wasm, so the same code ran with different widths. Its
five users are 32-bit quantities and now use uint32_t:

- ShiftDoubleReg() (SAD, the A:D register pair). It also shifted an int
  into bit 31 (msb << 31, tmp << 31, m << 31), which is undefined
  behaviour and, on native, sign-extended into bits 32-63 that the caller
  then discarded. A differential test of the old and new function over
  all 4 shift types x 64 count/sign encodings x both M values x 212
  register values (108,544 cases) found no difference in the result or
  in M; the new function is clean under UBSan, the old one is not.
- checkPageProtection() built its permit masks with 1L << 29..31, which
  overflows a 32-bit long on wasm; it now uses UINT32_C(1) << n.
- PhysMemRead/PhysMemWrite addresses and GetPageTableEntryDebugInfo().

The ulong typedef itself stays until the type cleanup step; nothing uses
it any more. The TPE instruction verifier has not been run on this
change.

+++ src/cpu/cpu.c
@@ -508,7 +508,7 @@ void device_interrupt(ushort interruptBits)
 /*
  * Routine that handles phys mem writes and shadow memory.
  */
-void PhysMemWrite(ushort value, ulong addr)
+void PhysMemWrite(ushort value, uint32_t addr)
 {
 	WritePhysicalMemory(addr, value, false); // in cpu_mms.c
 	return;
@@ -519,7 +519,7 @@ void PhysMemWrite(ushort value, ulong addr)
 /*
  * Routine that handles phys mem reads and shadow memory.
  */
-ushort PhysMemRead(ulong addr)
+ushort PhysMemRead(uint32_t addr)
 {
 	return ReadPhysicalMemory(addr, false); // in cpu_mms.c
